requirements.txt
# Core dependencies for the Chatbot with Mem0 and MCP Integration
streamlit==1.29.0
mem0ai==0.1.0
openai==1.54.3
httpx==0.27.0
python-dotenv==1.0.0
google-auth==2.25.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
requests==2.31.0
pandas==2.1.3
openpyxl==3.1.2
qdrant-client==1.7.0